Replace Provisioner by CloudAwareProvisioner
Since it's no use to have a provisioner which can't provision any VMs,
we may as well just have a base class instead of an interface.
